se modificó las clases

+++ tp2_grupo017/src/main/java/ar/edu/unju/fi/services/IServicioService.java
@@ -2,16 +2,16 @@ package ar.edu.unju.fi.services;
 
 import java.util.List;
 
-import ar.edu.unju.fi.models.Servicio;
+import ar.edu.unju.fi.entity.Servicio;
 
 public interface IServicioService {
     
     List<Servicio> getListado();
     Servicio getServicio();
-    Servicio getServicioBy(int id);
+    Servicio getServicioBy(Long id);
     public void guardarServicio(Servicio servcio);
     public void modificarServicio(Servicio servicioModificado);
-    public void eliminarServicio(int id);
+    public void eliminarServicio(Long id);
 
 
 }
